Initial commit.
Rough surface, be careful! This is just for fun, but you never know, it may turn real pretty soon.
Why? Because even in 2011 nothing beats a good old VT100 terminal app.

The idea is to create a framework that would work pretty much the same as jQuery UI / Mobile. Yeah!